+++ general/srcs/util.c
@@ -44,7 +44,7 @@ ssize_t		ft_arr_find(const char **str_arr, size_t size, const char *s)
 	i = 0;
 	if (!s)
 		return (-1);
-	while (i < size)
+	while ((size_t )i < size)
 	{
 		if (ft_strcmp(str_arr[i], s) == 0)
 			return (i);
